+++ packages/runtime-core/src/hmr.ts
@@ -160,11 +160,14 @@ function reload(id: string, newComp: HMRComponent): void {
         // don't end up forcing the same parent to re-render multiple times.
         queueJob(() => {
           isHmrUpdating = true
-          const parent = instance.parent!
+          const parent = instance.parent! as ComponentInternalInstance
           if (parent.vapor) {
             parent.hmrRerender!()
           } else {
-            ;(parent as ComponentInternalInstance).effect.run()
+            if (!(parent.effect.flags! & EffectFlags.STOP)) {
+              parent.renderCache = []
+              parent.effect.run()
+            }
           }
           nextTick(() => {
             isHmrUpdating = false